Remove PreBackupOperation and use BackupOperation instead
